Update - Add Unit Test for API Product Category - Update logic on parsing for product category

+++ fakestore_core_foundation/lib/models/fs_product.dart
@@ -14,7 +14,7 @@ class FSProduct extends TFModel {
   FSProduct.fromJson(JSONData? json) {
     id = json?["id"] ?? 0;
     title = json?["title"] ?? "";
-    price = json?["price"] ?? 0;
+    price = (json?["price"] ?? 0).toDouble();
     description = json?["description"] ?? "";
     image = json?["image"] ?? "";
     category = categoryEnumFromString(json?["category"] ?? "");
